Refactor: Remove compile time assertion flags
The idea here was to get some performance benefits by making some
assertions about the code, like validating that the property name is ok,
at compile time and then pass that info to the runtime function call
using a bit flag.

This brought significant complexity to the code and I never even tested
what, if any, actual performance benefits there would be. We can come
back to this approach leter if deemed necessary but it's way too early
to worry about microbenchmarks at this point when we don't even have a
POC yet.