Allow disabling Proxy wrapper for some objects
Not sure yet if this will be part of the public API or not. Right now
I'm just using this to make the code generated for tests look a bit
nicer. We don't need to generate

proxyfill$invoke(assert, 'strictEquals'

etc, as we know the assert calls work just fine without the verbose
wrappers.