Return polyfill from accessors too
The logic where we replace calls to polyfilled functions, such as
hasOwnProperty, was already implemented for `invoke` in previous commit.
However, we must also do that for `get` because an accessor could bind
it to a variable and call it directly, in which case the call won't go
through invoke.