Bail out of memberExpression when inside method call
It seems that at least in some cases Babel or babel-loader (not sure
which one's fault it is), will call the MemberExpression handler before
CallExpression for method call foo.bar(). Not quite sure what's
happening here exactly, I thought the order would be the same every
time, from outside to inside. Maybe some other Babel plugin manually
called traverse on the callee of the method?

Anyway, this fixes it by just cheking if we are the callee of a
CallExpression and doing early exit if that's the case.